+++ R/Rawls.R
@@ -27,7 +27,7 @@
 Rawls <-
     function()
 {
-    access_token <- .gcloud_access_token()
+    access_token <- .gcloud_access_token("rawls")
     api_header <- c(Authorization = paste("Bearer", access_token))
     .Rawls(
         Service(