+++ R/gcloud.R
@@ -5,19 +5,31 @@
 }
 
 .gcloud_access_token <-
-    function()
+    local(
 {
-    app_default <-
-        if (identical(Sys.getenv("USER"), "jupyter-user"))
-            "application-default"
-    .gcloud_do("auth", app_default, "print-access-token")
-}
+    tokens <- new.env(parent = emptyenv())
+    function(service) {
+        app_default <-
+            if (identical(Sys.getenv("USER"), "jupyter-user"))
+                "application-default"
+
+        key <- paste0(service, ":", app_default)
+        now <- Sys.time()
+
+        if (is.null(tokens[[key]]) || tokens[[key]]$expires < now) {
+            token <- .gcloud_do("auth", app_default, "print-access-token")
+            expires <- Sys.time() + 3600L
+            tokens[[key]] <- list(token = token, expires = expires)
+        }
+        tokens[[key]]$token
+    }
+})
 
 #' @rdname gcloud
 #'
 #' @name gcloud
 #'
-#' @title Interact with the gcloud command line utility
+#' @title gcloud command line utility interface
 #'
 #' @description These functions invoke the `gcloud` command line
 #'     utility. See \link{gsutil} for details on how `gcloud` is